# venture-forge/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Venture Forge Product API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FounderView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ventures": {
      "get": {
        "summary": "Ventures",
        "operationId": "ventures_api_v1_ventures_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentureList"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Venture",
        "operationId": "create_venture_api_v1_ventures_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VentureCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventures/{venture_id}/passport": {
      "get": {
        "summary": "Get Passport",
        "operationId": "get_passport_api_v1_ventures__venture_id__passport_get",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventures/{venture_id}/hypotheses": {
      "get": {
        "summary": "List Hypotheses",
        "operationId": "list_hypotheses_api_v1_ventures__venture_id__hypotheses_get",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HypothesisView"
                  },
                  "title": "Response List Hypotheses Api V1 Ventures  Venture Id  Hypotheses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Hypothesis",
        "operationId": "add_hypothesis_api_v1_ventures__venture_id__hypotheses_post",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "description",
          "created_at"
        ],
        "title": "ActivityView"
      },
...
      "FounderView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed"
          },
          "has_password": {
            "type": "boolean",
            "title": "Has Password"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "realm": {
            "type": "string",
            "const": "product",
            "title": "Realm",
            "default": "product"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "onboarding_completed",
          "has_password",
          "created_at"
        ],
        "title": "FounderView"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HypothesisCreate": {
        "properties": {
          "statement": {
            "type": "string",
            "maxLength": 1500,
            "minLength": 10,
            "title": "Statement"
          },
          "category": {
            "type": "string",
            "enum": [
              "problem",
              "customer",
              "pricing",
              "solution"
            ],
            "title": "Category",
            "default": "problem"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "statement",
          "expected_revision"
        ],
        "title": "HypothesisCreate"
      },
      "HypothesisView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "status": {
            "type": "string",
            "const": "unvalidated",
            "title": "Status"
          },
          "origin": {
            "type": "string",
            "const": "founder_asserted",
            "title": "Origin"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "statement",
          "category",
          "status",
          "origin",
          "revision",
          "created_at"
        ],
        "title": "HypothesisView"
      },
      "Login": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Login"
      },
...
      "Passport": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "realm": {
            "type": "string",
            "const": "product",
            "title": "Realm",
            "default": "product"
          },
          "venture": {
            "$ref": "#/components/schemas/VentureView"
          },
          "hypotheses": {
            "items": {
              "$ref": "#/components/schemas/HypothesisView"
            },
            "type": "array",
            "title": "Hypotheses"
          },
          "activity": {
            "items": {
              "$ref": "#/components/schemas/ActivityView"
            },
            "type": "array",
            "title": "Activity"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count",
            "default": 0
          },
          "decision_count": {
            "type": "integer",
            "title": "Decision Count",
            "default": 0
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "artifacts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "experiments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Experiments"
          },
          "decisions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Decisions"
          },
          "runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Runs"
          },
          "completed_cycles": {
            "type": "integer",
            "title": "Completed Cycles",
            "default": 0
          },
          "agent_runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Agent Runs"
          },
          "pipelines": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pipelines"
          },
          "handoffs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Handoffs"
          }
        },
        "type": "object",
        "required": [
          "venture",
          "hypotheses",
          "activity"
        ],
        "title": "Passport"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VentureCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "idea": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Idea"
          },
          "customer_segment": {
            "type": "string",
            "maxLength": 300,
            "minLength": 3,
            "title": "Customer Segment"
          },
          "geography": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Geography"
          },
          "first_hypothesis": {
            "type": "string",
            "maxLength": 1500,
            "minLength": 10,
            "title": "First Hypothesis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "idea",
          "customer_segment",
          "geography",
          "first_hypothesis"
        ],
        "title": "VentureCreate"
      },
      "VentureList": {
        "properties": {
          "ventures": {
            "items": {
              "$ref": "#/components/schemas/VentureView"
            },
            "type": "array",
            "title": "Ventures"
          }
        },
        "type": "object",
        "required": [
          "ventures"
        ],
        "title": "VentureList"
      },
      "VentureView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "idea": {
            "type": "string",
            "title": "Idea"
          },
          "customer_segment": {
            "type": "string",
            "title": "Customer Segment"
          },
          "geography": {
            "type": "string",
            "title": "Geography"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "idea",
          "customer_segment",
          "geography",
          "stage",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "VentureView"
      }
    }
  }
}